# docs/swagger.yaml
openapi: 3.0.3
info:
    title: UnaHur Anti-Social Net API
    description: |
        API REST del backend para la red social **UnaHur Anti-Social Net**.
        Permite gestionar usuarios, publicaciones, comentarios, etiquetas, imágenes asociadas a posts
        y relaciones de seguimiento entre usuarios (followers).

        **Variables de entorno relevantes**

        | Variable | Descripción |
        | --- | --- |
        | PORT | Puerto del servidor (por defecto 3001) |
        | MESES | Antigüedad máxima en meses para comentarios en publicaciones (por defecto 6) |
        | CACHE_POSTS_ENABLED | Caché en memoria para GET de posts (por defecto true) |
        | CACHE_POSTS_TTL_SECONDS | TTL del caché en segundos (por defecto 60) |
        | DB_DIALECT | sqlite, mysql, mariadb o postgres |
        | ENABLE_SWAGGER | Habilita Swagger UI en producción (por defecto false) |
    version: 1.0.0
    contact:
        name: Anti-Social Team
servers:
    - url: http://localhost:3001
      description: Servidor local (puerto configurable vía variable de entorno PORT)
...
paths:
    /users:
        get:
            tags: [Users]
            summary: Listar todos los usuarios
            operationId: getAllUsers
            parameters:
                - $ref: "#/components/parameters/ViewerIdQuery"
            responses:
                "200":
                    description: |
                        Lista de usuarios. Perfiles privados devuelven solo `id`, `nickname` e `isProfilePublic`
                        salvo que `viewer_id` sea el propio usuario o un seguidor.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    oneOf:
                                        - $ref: "#/components/schemas/User"
                                        - $ref: "#/components/schemas/UserLimited"
        post:
            tags: [Users]
            summary: Crear un usuario
            operationId: createUser
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UserCreate"
            responses:
                "201":
                    description: Usuario creado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
                "400":
                    $ref: "#/components/responses/ValidationError"
                "409":
                    $ref: "#/components/responses/ConflictError"
    /users/{id}:
        parameters:
            - $ref: "#/components/parameters/IdParam"
        get:
            tags: [Users]
            summary: Obtener un usuario por ID
            operationId: getUserById
            parameters:
                - $ref: "#/components/parameters/ViewerIdQuery"
            responses:
                "200":
                    description: Usuario encontrado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "403":
                    $ref: "#/components/responses/Forbidden"
                "404":
                    $ref: "#/components/responses/NotFound"
        put:
            tags: [Users]
            summary: Actualizar un usuario
            operationId: updateUser
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UserUpdate"
            responses:
                "200":
                    description: Usuario actualizado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
                "400":
                    $ref: "#/components/responses/ValidationError"
                "404":
                    $ref: "#/components/responses/NotFound"
                "409":
                    $ref: "#/components/responses/ConflictError"
        delete:
            tags: [Users]
            summary: Eliminar un usuario
            operationId: deleteUser
            responses:
                "200":
                    description: Usuario eliminado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/MessageResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
    /users/{id}/profile-picture:
        parameters:
            - $ref: "#/components/parameters/IdParam"
        put:
            tags: [Users]
            summary: Subir o actualizar la foto de perfil
            operationId: uploadUserProfilePicture
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            required: [image]
                            properties:
                                image:
                                    type: string
                                    format: binary
                                    description: Archivo de imagen (JPEG, PNG o WebP, máx. 5 MB)
            responses:
                "200":
                    description: Foto de perfil actualizada
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
                "400":
                    $ref: "#/components/responses/ValidationError"
                "404":
                    $ref: "#/components/responses/NotFound"
        delete:
            tags: [Users]
            summary: Eliminar la foto de perfil
            operationId: deleteUserProfilePicture
            responses:
                "200":
                    description: Foto de perfil eliminada
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/MessageResponse"
                "404":
                    $ref: "#/components/responses/NotFound"
    /users/{id}/followers:
        parameters:
            - $ref: "#/components/parameters/IdParam"
        get:
            tags: [Followers]
            summary: Listar seguidores de un usuario
            description: Usuarios que siguen al usuario indicado por `id`.
            operationId: getUserFollowers
            parameters:
                - $ref: "#/components/parameters/ViewerIdQuery"
            responses:
                "200":
                    description: Lista de seguidores
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/UserSummary"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "403":
                    $ref: "#/components/responses/Forbidden"
                "404":
                    $ref: "#/components/responses/NotFound"
    /users/{id}/following:
        parameters:
            - $ref: "#/components/parameters/IdParam"
        get:
            tags: [Followers]
            summary: Listar usuarios seguidos
            description: Usuarios que el usuario `id` está siguiendo.
            operationId: getUserFollowing
            parameters:
                - $ref: "#/components/parameters/ViewerIdQuery"
            responses:
                "200":
                    description: Lista de usuarios seguidos
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/UserSummary"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "403":
                    $ref: "#/components/responses/Forbidden"
                "404":
                    $ref: "#/components/responses/NotFound"
    /users/{id}/follow:
        parameters:
            - $ref: "#/components/parameters/IdParam"
        post:
            tags: [Followers]
            summary: Seguir a un usuario
            description: |
                El usuario `follower_id` comienza a seguir al usuario `id`.
                No se permite seguirse a uno mismo.
            operationId: followUser
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/FollowRequest"
            responses:
                "201":
                    description: Seguimiento creado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/FollowActionResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
                "409":
                    $ref: "#/components/responses/ConflictError"
    /users/{id}/follow/{follower_id}:
        parameters:
            - $ref: "#/components/parameters/IdParam"
            - name: follower_id
              in: path
              required: true
              schema:
                  type: integer
              description: ID del usuario que deja de seguir
        delete:
            tags: [Followers]
            summary: Dejar de seguir a un usuario
            operationId: unfollowUser
            responses:
                "200":
                    description: Seguimiento eliminado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/FollowActionResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
...
    /tags:
        get:
            tags: [Tags]
            summary: Listar todas las etiquetas
            description: |
                Devuelve etiquetas con las publicaciones asociadas.
                Opcionalmente filtra por `post_id` (solo tags vinculados a esa publicación).
            operationId: getAllTags
            parameters:
                - name: post_id
                  in: query
                  description: Filtrar etiquetas asociadas a una publicación
                  required: false
                  schema:
                      type: integer
                      minimum: 1
            responses:
                "200":
                    description: Lista de tags
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Tag"
                "400":
                    $ref: "#/components/responses/ValidationError"
                "404":
                    $ref: "#/components/responses/NotFound"
        post:
            tags: [Tags]
            summary: Crear una etiqueta
            description: |
                Crea una etiqueta por name o reutiliza la existente (mismo nombre, sin distinguir mayúsculas).
                Si se envía post_id, la vincula a esa publicación.
                Responde 201 si el tag es nuevo; 200 si ya existía (en ambos casos devuelve el tag con sus posts).
            operationId: createTag
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/TagCreate"
            responses:
                "200":
                    description: Tag ya existía; se devuelve el existente (y se vincula al post si se envió post_id)
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Tag"
                "201":
                    description: Tag creado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Tag"
                "400":
                    $ref: "#/components/responses/ValidationError"
                "404":
                    $ref: "#/components/responses/NotFound"
    /tags/{id}:
        parameters:
            - $ref: "#/components/parameters/IdParam"
        get:
            tags: [Tags]
            summary: Obtener una etiqueta por ID
            operationId: getTagById
            responses:
                "200":
                    description: Tag encontrado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Tag"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
        patch:
            tags: [Tags]
            summary: Actualizar una etiqueta
            operationId: updateTag
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/TagUpdate"
            responses:
                "200":
                    description: Tag actualizado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Tag"
                "400":
                    $ref: "#/components/responses/ValidationError"
                "404":
                    $ref: "#/components/responses/NotFound"
        delete:
            tags: [Tags]
            summary: Eliminar una etiqueta
            operationId: deleteTag
            responses:
                "200":
                    description: Tag eliminado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/MessageResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
components:
    parameters:
        IdParam:
            name: id
            in: path
            required: true
            description: Identificador numérico del recurso
            schema:
                type: integer
                minimum: 1
...
        ViewerIdQuery:
            name: viewer_id
            in: query
            required: false
            description: ID del usuario que consulta el perfil (para validar perfiles privados)
            schema:
                type: integer
                minimum: 1
    responses:
        BadRequest:
            description: Solicitud inválida
            content:
                application/json:
                    schema:
                        oneOf:
                            - $ref: "#/components/schemas/MessageResponse"
                            - $ref: "#/components/schemas/ValidationErrorResponse"
        ValidationError:
            description: Error de validación de datos
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ValidationErrorResponse"
        NotFound:
            description: Recurso no encontrado
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/MessageResponse"
        Forbidden:
            description: Acceso denegado al perfil privado
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/MessageResponse"
        ConflictError:
            description: Conflicto por restricción de unicidad
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
    schemas:
        MessageResponse:
            type: object
            properties:
                message:
                    type: string
                    example: Operación realizada correctamente
        ErrorResponse:
            type: object
            properties:
                message:
                    type: string
                error:
                    type: string
        ValidationErrorResponse:
            type: object
            properties:
                errores:
                    type: array
                    items:
                        type: object
                        properties:
                            atributo:
                                type: string
                            error:
                                type: string
        User:
            type: object
            properties:
                id:
                    type: integer
                    example: 1
                nickname:
                    type: string
                    example: juan_perez
                name:
                    type: string
                    example: Juan
                lastName:
                    type: string
                    example: Perez
                email:
                    type: string
                    format: email
                    example: juan@example.com
                password:
                    type: string
                    writeOnly: true
                birthDate:
                    type: string
                    format: date-time
                gender:
                    type: string
                    enum: [femenino, Femenino, masculino, Masculino, otro, Otro]
                profilePicture:
                    type: string
                    format: uri
                    nullable: true
                    example: http://localhost:3001/uploads/profiles/foto.jpg
                isProfilePublic:
                    type: boolean
                    description: true = perfil público; false = solo visible para seguidores
                    example: true
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
        UserCreate:
            type: object
            required: [nickname, name, lastName, email, password, birthDate, gender]
            properties:
                nickname:
                    type: string
                    minLength: 3
                    maxLength: 30
                    pattern: "^[a-zA-Z0-9_.-]+$"
                name:
                    type: string
                    minLength: 2
                    maxLength: 100
                lastName:
                    type: string
                    minLength: 2
                    maxLength: 100
                email:
                    type: string
                    format: email
                password:
                    type: string
                    minLength: 6
                    pattern: "^[a-zA-Z0-9_.-]+$"
                birthDate:
                    type: string
                    format: date
                gender:
                    type: string
                    enum: [femenino, Femenino, masculino, Masculino, otro, Otro]
                isProfilePublic:
                    type: boolean
                    default: true
                    description: true = perfil público; false = solo visible para seguidores
        UserUpdate:
            type: object
            minProperties: 1
            properties:
                nickname:
                    type: string
                    minLength: 3
                    maxLength: 30
                name:
                    type: string
                    minLength: 2
                    maxLength: 100
                lastName:
                    type: string
                    minLength: 2
                    maxLength: 100
                email:
                    type: string
                    format: email
                password:
                    type: string
                    minLength: 6
                birthDate:
                    type: string
                    format: date
                gender:
                    type: string
                    enum: [femenino, Femenino, masculino, Masculino, otro, Otro]
                isProfilePublic:
                    type: boolean
                    description: true = perfil público; false = solo visible para seguidores
        UserLimited:
            type: object
            description: Vista reducida de un perfil privado
            properties:
                id:
                    type: integer
                nickname:
                    type: string
                isProfilePublic:
                    type: boolean
                    example: false
...
        UserSummary:
            type: object
            properties:
                id:
                    type: integer
                nickname:
                    type: string
                name:
                    type: string
                lastName:
                    type: string
                profilePicture:
                    type: string
                    format: uri
                    nullable: true
...
        PostSummary:
            type: object
            properties:
                id:
                    type: integer
                description:
                    type: string
...
        Tag:
            type: object
            properties:
                id:
                    type: integer
                name:
                    type: string
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
                posts:
                    type: array
                    description: Publicaciones asociadas a la etiqueta
                    items:
                        $ref: "#/components/schemas/PostSummary"
        TagCreate:
            type: object
            required: [name]
            properties:
                name:
                    type: string
                    minLength: 1
                    maxLength: 25
                post_id:
                    type: integer
                    minimum: 1
                    description: Publicación a la que asociar la etiqueta al crearla (opcional)
        TagUpdate:
            type: object
            minProperties: 1
            properties:
                name:
                    type: string
                    minLength: 1
                    maxLength: 25
        FollowRequest:
            type: object
            required: [follower_id]
            properties:
                follower_id:
                    type: integer
                    minimum: 1
                    description: ID del usuario que realiza o deja el seguimiento
        FollowActionResponse:
            type: object
            properties:
                message:
                    type: string
                    example: Seguimiento creado correctamente
                following_id:
                    type: integer
                    description: ID del usuario seguido
                follower_id:
                    type: integer
                    description: ID del usuario que sigue